moved product types ctl to product ctrl and changed route path to reflect change
