+++ routes/index.js
@@ -9,7 +9,7 @@ router.get('/', (req, res, next) => {
 
 // pipe all other requests through the route modules
 router.use(require('./authRoute'));
-router.use(require('./productTypeRoute'));
+router.use(require('./productRoute'));
 
 // router.use(require('./foo'));
