+++ routes/productRoute.js
@@ -3,7 +3,7 @@
 const { Router } = require('express');
 const router = Router();
 
-const { getProductTypes } = require('../controllers/productTypeCtrl.js');
+const { getProductTypes } = require('../controllers/productCtrl.js');
 
 router.get('/prodTypes', getProductTypes);
